Reorganize according to new structure
