+++ pymtl3/passes/sverilog/errors.py
