+++ pymtl3/passes/sverilog/translation/behavioral/SVBehavioralTranslatorL0.py
@@ -7,7 +7,7 @@
 
 from __future__ import absolute_import, division, print_function
 
-from pymtl.passes.translator.behavioral.BehavioralTranslatorL0 import (
+from pymtl3.passes.translator.behavioral.BehavioralTranslatorL0 import (
     BehavioralTranslatorL0,
 )
