+++ pymtl3/passes/sverilog/translation/behavioral/SVBehavioralTranslatorL2.py
@@ -6,10 +6,10 @@
 """Provide the level 2 SystemVerilog translator implementation."""
 from __future__ import absolute_import, division, print_function
 
-from pymtl.passes.rtlir import BehavioralRTLIR as bir
-from pymtl.passes.rtlir import RTLIRType as rt
-from pymtl.passes.sverilog.utility import make_indent
-from pymtl.passes.translator.behavioral.BehavioralTranslatorL2 import (
+from pymtl3.passes.rtlir import BehavioralRTLIR as bir
+from pymtl3.passes.rtlir import RTLIRType as rt
+from pymtl3.passes.sverilog.utility import make_indent
+from pymtl3.passes.translator.behavioral.BehavioralTranslatorL2 import (
     BehavioralTranslatorL2,
 )
