+++ pymtl3/passes/sverilog/translation/behavioral/test/SVBehavioralTranslatorL1_test.py
@@ -7,11 +7,12 @@
 
 from __future__ import absolute_import, division, print_function
 
-from pymtl import Bits1, Bits32, Bits64, Component, InPort, OutPort, concat, sext, zext
-from pymtl.passes.rtlir import BehavioralRTLIRGenPass, BehavioralRTLIRTypeCheckPass
-from pymtl.passes.rtlir.test.test_utility import do_test, expected_failure
-from pymtl.passes.sverilog.errors import SVerilogTranslationError
-from pymtl.passes.sverilog.translation.behavioral.SVBehavioralTranslatorL1 import (
+from pymtl3.datatypes import Bits1, Bits32, Bits64, concat, sext, zext
+from pymtl3.dsl import Component, InPort, OutPort
+from pymtl3.passes.rtlir import BehavioralRTLIRGenPass, BehavioralRTLIRTypeCheckPass
+from pymtl3.passes.rtlir.util.test_utility import do_test, expected_failure
+from pymtl3.passes.sverilog.errors import SVerilogTranslationError
+from pymtl3.passes.sverilog.translation.behavioral.SVBehavioralTranslatorL1 import (
     BehavioralRTLIRToSVVisitorL1,
 )
