+++ pymtl3/passes/sverilog/translation/structural/SVStructuralTranslatorL2.py
@@ -6,9 +6,9 @@ from __future__ import absolute_import, division, print_function
 
 from functools import reduce
 
-from pymtl.passes.rtlir import RTLIRDataType as rdt
-from pymtl.passes.sverilog.utility import make_indent
-from pymtl.passes.translator.structural.StructuralTranslatorL2 import (
+from pymtl3.passes.rtlir import RTLIRDataType as rdt
+from pymtl3.passes.sverilog.utility import make_indent
+from pymtl3.passes.translator.structural.StructuralTranslatorL2 import (
     StructuralTranslatorL2,
 )
