+++ pymtl3/passes/sverilog/translation/structural/SVStructuralTranslatorL4.py
@@ -6,10 +6,10 @@ from __future__ import absolute_import, division, print_function
 
 from functools import reduce
 
-from pymtl.passes.rtlir import RTLIRDataType as rdt
-from pymtl.passes.rtlir import RTLIRType as rt
-from pymtl.passes.sverilog.utility import make_indent
-from pymtl.passes.translator.structural.StructuralTranslatorL4 import (
+from pymtl3.passes.rtlir import RTLIRDataType as rdt
+from pymtl3.passes.rtlir import RTLIRType as rt
+from pymtl3.passes.sverilog.utility import make_indent
+from pymtl3.passes.translator.structural.StructuralTranslatorL4 import (
     StructuralTranslatorL4,
 )
