+++ pymtl3/passes/sverilog/translation/structural/test/SVStructuralTranslatorL1_test.py
@@ -7,9 +7,10 @@
 
 from __future__ import absolute_import, division, print_function
 
-from pymtl import Bits1, Bits4, Bits32, Component, InPort, OutPort, Wire
-from pymtl.passes.rtlir.test.test_utility import do_test
-from pymtl.passes.sverilog.translation.structural.SVStructuralTranslatorL1 import (
+from pymtl3.datatypes import Bits1, Bits4, Bits32
+from pymtl3.dsl import Component, InPort, OutPort, Wire
+from pymtl3.passes.rtlir.util.test_utility import do_test
+from pymtl3.passes.sverilog.translation.structural.SVStructuralTranslatorL1 import (
     SVStructuralTranslatorL1,
 )
