+++ pymtl3/passes/sverilog/translation/structural/test/SVStructuralTranslatorL2_test.py
@@ -7,10 +7,11 @@
 
 from __future__ import absolute_import, division, print_function
 
-from pymtl import Bits1, Bits32, Component, InPort, OutPort, Wire
-from pymtl.passes.rtlir import RTLIRDataType as rdt
-from pymtl.passes.rtlir.test.test_utility import do_test
-from pymtl.passes.sverilog.translation.structural.SVStructuralTranslatorL2 import (
+from pymtl3.datatypes import Bits1, Bits32
+from pymtl3.dsl import Component, InPort, OutPort, Wire
+from pymtl3.passes.rtlir import RTLIRDataType as rdt
+from pymtl3.passes.rtlir.util.test_utility import do_test
+from pymtl3.passes.sverilog.translation.structural.SVStructuralTranslatorL2 import (
     SVStructuralTranslatorL2,
 )
